prepping for audio playback
git-svn-id: http://mediamods.com/public-svn/camera-activity/Record.activity@366 574bc980-5f2d-0410-acbc-c8f9f0eb14e0

+++ RecordActivity.py
@@ -44,12 +44,16 @@ class RecordActivity(activity.Activity):
 		activity.Activity.__init__(self, handle)
 		self.activityName = "Record"
 		self.set_title( self.activityName )
+
 		#wait a moment so that our debug console capture mistakes
 		gobject.idle_add( self._initme, None )
 
 	def _initme( self, userdata=None ):
 		self.instanceId = self._activity_id
 		self.ACTIVE = True
+
+		self.I_AM_CLOSING = False
+		self.I_AM_SAVED = False
 
 		self.nickName = profile.get_nick_name()
 		self.basePath = activity.get_bundle_path()
@@ -84,7 +88,6 @@ class RecordActivity(activity.Activity):
 
 		#listen for meshins
 		self.connect( "shared", self._sharedCb )
-		self.connect( "destroy", self._destroyCb)
 
 		#todo: proper focus listeners to turn camera on / off
 		#self.connect("focus-in-event", self.c.inFocus)
@@ -116,13 +119,190 @@ class RecordActivity(activity.Activity):
 
 
 	def write_file(self, file):
-		#todo: thread this, and the delete method called after it
 		print("write_file")
+		self.I_AM_SAVED = False
+		SAVING_AT_LEAST_ONE = False
+
 		#todo: just pass the file over to the method in m
 		f = open( file, "w" )
-		album = self.m.updateMediaIndex( )
+
+		print("updateMediaIndex")
+		impl = getDOMImplementation()
+		album = impl.createDocument(None, "album", None)
+		root = album.documentElement
+
+		photoHash = self.m.mediaHashs[self.m.TYPE_PHOTO]
+		for i in range (0, len(photoHash)):
+			recd = photoHash[i]
+
+			photo = album.createElement('photo')
+			root.appendChild(photo)
+			savingFile = self.saveMedia( photo, recd, self.m.TYPE_PHOTO )
+			if (savingFile):
+				SAVING_AT_LEAST_ONE = True
+
+		videoHash = self.m.mediaHashs[self.m.TYPE_VIDEO]
+		for i in range (0, len(videoHash)):
+			recd = videoHash[i]
+
+			video = album.createElement('video')
+			root.appendChild(video)
+			savingFile = self.saveMedia( video, recd, self.m.TYPE_VIDEO )
+			if (savingFile):
+				SAVING_AT_LEAST_ONE = True
+
+		audioHash = self.mediaHashs[self.m.TYPE_AUDIO]
+		for i in range (0, len(audioHash)):
+			recd = audioHash[i]
+
+			audio = album.createElement('audio')
+			root.appendChild(audio)
+			saving_file = self.saveMedia( audio, recd, self.m.TYPE_AUDIO )
+			if (saving_file):
+				SAVING_AT_LEAST_ONE = True
+
+
 		album.writexml(f)
 		f.close()
+
+		#if (SAVING_AT_LEAST_ONE):
+		#	self.I_AM_SAVED = False
+		#else:
+		#	self.I_AM_SAVED = True
+		self.I_AM_SAVED = not SAVING_AT_LEAST_ONE
+
+	def saveMedia( self, el, recd, type ):
+		needToDatastoreMedia = False
+
+		if ( (recd.buddy == True) and (recd.datastoreId == None) and (not recd.downloadedFromBuddy) ):
+			pixbuf = recd.getThumbPixbuf( )
+			buddyThumb = str( self._get_base64_pixbuf_data(pixbuf) )
+			el.setAttribute("buddyThumb", buddyThumb )
+		else:
+			needToDatastoreMedia = self.saveMediaToDatastore( recd )
+
+		el.setAttribute("type", str(type))
+		el.setAttribute("title", recd.title)
+		el.setAttribute("time", str(recd.time))
+		el.setAttribute("photographer", recd.photographer)
+		el.setAttribute("mediaFilename", recd.mediaFilename)
+		el.setAttribute("thumbFilename", recd.thumbFilename)
+		el.setAttribute("colorStroke", str(recd.colorStroke.hex) )
+		el.setAttribute("colorFill", str(recd.colorFill.hex) )
+		el.setAttribute("hashKey", str(recd.hashKey))
+		el.setAttribute("buddy", str(recd.buddy))
+		el.setAttribute("mediaMd5", str(recd.mediaMd5))
+		el.setAttribute("thumbMd5", str(recd.thumbMd5))
+		if (recd.datastoreId != None):
+			el.setAttribute("datastoreId", str(recd.datastoreId))
+
+		return needToDatastoreMedia
+
+
+	def saveMediaToDatastore( self, recd ):
+		#TODO: update the recds that go through here to how they would look on a fresh load from file since this won't just happen on close()
+
+		if (recd.datastoreId != None):
+			#already saved to the datastore, don't need to re-rewrite the file since the mediums are immutable
+			#However, they might have changed the name of the file
+			if (recd.titleChange):
+				self.loadMediaFromDatastore( recd )
+				try:
+					if (recd.datastoreOb.metadata['title'] != recd.title):
+						recd.datastoreOb.metadata['title'] = recd.title
+						datastore.write(recd.datastoreOb)
+						if (recd.datastoreOb != None):
+							recd.datastoreOb.destroy()
+							del recd.datastoreOb
+				finally:
+					if (recd.datastoreOb != None):
+						recd.datastoreOb.destroy()
+						del recd.datastoreOb
+
+				#reset for the next title change if not closing...
+				recd.titleChange = False
+			return False
+
+		#this will remove the media from being accessed on the local disk since it puts it away into cold storage
+		#therefore this is only called when write_file is called by the activity superclass
+		try:
+			mediaObject = datastore.create()
+			try:
+				#todo: what other metadata to set?
+				mediaObject.metadata['title'] = recd.title
+				#jobject.metadata['keep'] = '0'
+				#jobject.metadata['buddies'] = ''
+
+				pixbuf = recd.getThumbPixbuf()
+				thumbData = self._get_base64_pixbuf_data(pixbuf)
+				mediaObject.metadata['preview'] = thumbData
+
+				colors = str(recd.colorStroke.hex) + "," + str(recd.colorFill.hex)
+				mediaObject.metadata['icon-color'] = colors
+
+				if (recd.type == self.TYPE_PHOTO):
+					mediaObject.metadata['mime_type'] = 'image/jpeg'
+				elif (recd.type == self.TYPE_VIDEO):
+					mediaObject.metadata['mime_type'] = 'video/ogg'
+				elif (recd.type == self.TYPE_AUDIO):
+					mediaObject.metadata['mime_type'] = 'audio/ogg'
+
+				mediaFile = os.path.join(self.ca.journalPath, recd.mediaFilename)
+				mediaObject.file_path = mediaFile
+
+				datastore.write(mediaObject, reply_handler=self._mediaSaveCb, error_handler=self._mediaSaveErrorCb )
+				#todo: verify that object_id is set even with callbacks
+				print( "datastoreId: ", mediaObject.object_id )
+				recd.datastoreId = mediaObject.object_id
+
+			finally:
+				if (mediaObject != None):
+					mediaObject.destroy()
+					del mediaObject
+					return True
+				else:
+					return False
+
+		finally:
+			#unable to create a mediaObject, so nothing to save, return False
+			return False
+			#todo: don't really need to do this here, since we delete our temp before shutdown, but what about other write_files?
+			#os.remove(file_path)
+
+
+	def _mediaSaveCb( self ):
+		self.doPostMediaSave()
+
+
+	def _mediaSaveErrorCb( self ):
+		self.doPostMediaSave()
+
+
+	def doPostMediaSave( self ):
+		allDone = True
+
+		photoHash = self.m.mediaHashs[self.m.TYPE_PHOTO]
+		for i in range (0, len(photoHash)):
+			recd = photoHash[i]
+			if (recd.dirty):
+				allDone = False
+
+		videoHash = self.m.mediaHashs[self.m.TYPE_VIDEO]
+		for i in range (0, len(videoHash)):
+			recd = videoHash[i]
+			if (recd.dirty):
+				allDone = False
+
+		audioHash = self.mediaHashs[self.m.TYPE_AUDIO]
+		for i in range (0, len(audioHash)):
+			recd = audioHash[i]
+			if (recd.dirty):
+				allDone = False
+
+		if (allDone):
+			self.I_AM_SAVED = True
+		if (self.I_AM_SAVED and self.I_AM_CLOSING):
+			self.destroy()
 
 
 	def _sharedCb( self, activity ):
@@ -160,13 +340,14 @@ class RecordActivity(activity.Activity):
 
 
 	def stopPipes(self):
-		#todo: if recording a movie when you leave, stop.  also make sure not to put the video back on display when done.
+		#todo: also make sure not to put the video back on display when done.
 		print("stop pipes")
 		self.gplay.stop()
 		self.ui.doMouseListener( False )
 
 		if (self.m.RECORDING):
-			self.m.stopRecordingVideo()
+			self.m.setUpdating( False )
+			self.m.doShutter()
 		else:
 			self.glive.stop()
 
@@ -187,11 +368,10 @@ class RecordActivity(activity.Activity):
 
 	def close( self ):
 		print("close")
+		self.I_AM_CLOSING = True
 		self.m.UPDATING = False
 		self.ui.updateButtonSensitivities( )
-
 		self.ui.doMouseListener( False )
-
 		self.ui.hideLiveWindows( )
 		self.ui.hidePlayWindows( )
 		self.ui.audioCanvas.stopWaveformDraws( )
@@ -200,21 +380,23 @@ class RecordActivity(activity.Activity):
 		self.glive.stopAudioHandoffs( )
 		self.glive.stop( )
 
-
-		#todo: put this into a gobject.freetime thing to update the ui first?
 		activity.Activity.close( self )
-		#this deletes all the temp files after the call above saves them...
-		self._destroyCb( None )
 
-	def _destroyCb( self, *args ):
-		#todo: disable the buttons while this is going down
-		print("destroyCb")
 
-		#todo: why recreate temp and destroy journalpath?
-		#todo: clean up / throw away any video you might be recording when you quit the activity
-		self.recreateTemp()
+	def destroy( self ):
+		print( "destroy and I_AM_CLO", self.I_AM_CLOSING, "sav..", self.I_AM_SAVED )
 
-		if (os.path.exists(self.journalPath)):
-			shutil.rmtree( self.journalPath )
+		if self.I_AM_CLOSING:
+			self.hide()
 
-		gtk.main_quit()
+		if self.I_AM_SAVED:
+			print("total Destruction")
+
+			#todo: why recreate temp and destroy journalpath?
+			#todo: clean up / throw away any video you might be recording when you quit the activity
+			self.recreateTemp()
+
+			if (os.path.exists(self.journalPath)):
+				shutil.rmtree( self.journalPath )
+
+			activity.Activity.destroy()
